# src/document/swagger-final.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sekolah Kopi Raisa",
    "version": "1.0.0",
    "description": "API untuk aplikasi Sekolah Kopi Raisa - E-commerce dan Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://be-web-sekolah-kopi-raisa.vercel.app",
      "description": "Vercel Production Server"
    },
    {
      "url": "https://sekolah-kopi-raisa.up.railway.app",
      "description": "Railway Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  },
  "paths": {
...
    "/api/v1/order": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get all orders (admin)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create new order",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array"
                  },
                  "shippingAddress": {
                    "type": "string"
                  },
                  "paymentMethod": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully"
          }
        }
      }
    },
    "/api/v1/order/{id}/detail": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order detail",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order detail retrieved successfully"
          }
        }
      }
    },
    "/api/v1/order/my-order": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get user's orders",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User orders retrieved successfully"
          }
        }
      }
    },
    "/api/v1/order/history": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order history",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Order history retrieved successfully"
          }
        }
      }
    },
    "/api/v1/order/completed": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get completed orders",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Completed orders retrieved successfully"
          }
        }
      }
    },
    "/api/v1/order/Order-statuses": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order statuses",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Order statuses retrieved successfully"
          }
        }
      }
    },
    "/api/v1/order/payment-method": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get payment methods",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payment methods retrieved successfully"
          }
        }
      }
    },
...
    "/api/v1/order/contact-partner/{partnerId}": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Contact partner for order",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "partnerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Partner contacted successfully"
          }
        }
      }
    },
    "/api/v1/order/search-address": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Search shipping address",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Address search completed successfully"
          }
        }
      }
    },
    "/api/v1/order/search-cost": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Calculate shipping cost",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Shipping cost calculated successfully"
          }
        }
      }
    },
    "/api/v1/order/{id}/cancel": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Cancel order",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled successfully"
          }
        }
      }
    },
    "/api/v1/order/{id}/update-status": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Update order status",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order status updated successfully"
          }
        }
      }
    },
    "/api/v1/order/{id}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Update order",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Delete order",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully"
          }
        }
      }
    },
    "/api/v1/order/notifications": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order notifications",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully"
          }
        }
      }
    },
...
  }
}